Move components used in many pages to common folder
